+++ code/system.ui/react.common/src/ui/EdgePosition.Selector/view.Cell.tsx
